tools/cloudops-copilot/requirements.txt
aiogram>=3.20.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
aiohttp>=3.11.0,<4.0.0
requests>=2.32.0,<3.0.0
azure-identity>=1.17.0,<2.0.0
azure-monitor-query>=1.4.0,<2.0.0
azure-mgmt-monitor>=6.0.2,<7.0.0
azure-mgmt-compute>=34.0.0,<35.0.0
azure-mgmt-sql>=3.0.1,<4.0.0
azure-mgmt-network>=28.1.0,<29.0.0
PyYAML>=6.0.2,<7.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.8.0,<3.0.0
